requirements.txt
# python==3.10.5
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
psycopg2==2.9.5
python-dotenv==0.21.1
Flask-Bcrypt==1.0.1
WTForms==3.0.1
email-validator==1.3.1
Flask-Mail==0.9.1
PyJWT==2.6.0
Flask-Migrate==4.0.4
numpy==1.24.1
pandas==1.5.3
scikit-learn==1.0.2